requirements.txt
asn1crypto==1.5.1
bottle==0.12.23
gunicorn==20.1.0
pg8000==1.29.4
PyMySQL==0.9.3
python-dateutil==2.8.2
python-dotenv==0.21.0
scramp==1.4.4
six==1.16.0
